chore: extract fetch hooks for common table
